+++ src/locales/es/messages.po
@@ -472,7 +472,7 @@ msgstr "Tasa de Préstamo / Día"
 msgid "Borrow Fee Rate"
 msgstr ""
 
-#: src/domain/synthetics/markets/getBridigingInfo.tsx
+#: src/components/Synthetics/BridgingInfo/BridgingInfo.tsx
 msgid "Bridge {tokenSymbol} to {chainName} using any of the options below:"
 msgstr ""
